memea
Enum MemeaError

pub enum MemeaError {

    GdsParse(GdsError),
    Gds(GdsError),
    Io(Error),
    ParseInt(ParseIntError),
    ParseFloat(ParseFloatError),
    Config(ConfigError),
    Lef(LefError),
    Dialogue(Error),
    SerdeYaml(Error),
    SerdeJson(Error),
    CSV(Error),
    ParseError(String),
    DatabaseError(DBError),

}

Comprehensive error type for all MemEA operations.

This enum covers all possible errors that can occur during MemEA operations, including file I/O, parsing, database operations, and user interaction errors. Most variants automatically convert from their underlying error types using the #[from] attribute.

Variants

GdsParse(GdsError)

GDS parsing-specific error from the gds module.

Gds(GdsError)

GDS library error from the gds21 crate.

Io(Error)

Standard I/O error (file operations, etc.).

ParseInt(ParseIntError)

Integer parsing error.

ParseFloat(ParseFloatError)

Floating-point parsing error.

Config(ConfigError)

Configuration file parsing error.

Lef(LefError)

LEF file parsing error.

Dialogue(Error)

User interaction error from dialoguer.

SerdeYaml(Error)

YAML serialization/deserialization error.

SerdeJson(Error)

JSON serialization/deserialization error.

CSV(Error)

CSV export error.

ParseError(String)

General parsing error with custom message.

DatabaseError(DBError)

Database operation error.